+++ packages/rulesets/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@stoplight/spectral-rulesets-v1.16.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.15.1...@stoplight/spectral-rulesets-v1.16.0) (2023-04-25)
+
+
+### Features
+
+* **rulesets:** add traits array path to headers rule ([#2460](https://github.com/stoplightio/spectral/issues/2460)) ([9ceabca](https://github.com/stoplightio/spectral/commit/9ceabca80969885c240349d6ebba15c09a4f8697))
+
+# [@stoplight/spectral-rulesets-v1.15.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.15.0...@stoplight/spectral-rulesets-v1.15.1) (2023-04-25)
+
+
+### Bug Fixes
+
+* **rulesets:** avoid false errors from ajv ([#2408](https://github.com/stoplightio/spectral/issues/2408)) ([92dab78](https://github.com/stoplightio/spectral/commit/92dab78d0c07e6919c0485cadbe5aa2391a53e8b))
+
 # [@stoplight/spectral-rulesets-v1.15.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.14.1...@stoplight/spectral-rulesets-v1.15.0) (2023-02-03)